elaim-bot/requirements.txt
discord.py==2.3.2
pydantic==2.5.2
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
